+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/controller/PlanController.java
@@ -32,7 +32,7 @@ public class PlanController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<PlanResponse> porId(@PathVariable Long id) {
+    public ResponseEntity<PlanResponse> porId(@PathVariable Integer id) {
         return ResponseEntity.ok(service.findById(id));
     }
 
@@ -43,13 +43,13 @@ public class PlanController {
 
     @PutMapping("/{id}")
     public ResponseEntity<PlanResponse> actualizar(
-            @PathVariable Long id,
+            @PathVariable Integer id,
             @Valid @RequestBody PlanRequest request) {
         return ResponseEntity.ok(service.save(request, id));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> eliminar(@PathVariable Long id) {
+    public ResponseEntity<Void> eliminar(@PathVariable Integer id) {
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
