+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/dto/request/CarritoRequest.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class CarritoRequest {
     @NotBlank
     private String sessionId;
-    private Long varianteId;
+    private Integer varianteId;
     @Positive
     private Integer cantidad = 1;
 }
