+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/dto/response/TiendaResponse.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class TiendaResponse {
-    private Long id;
+    private Integer id;
     private String nombre;
     private String slug;
     private String whatsapp;
@@ -12,8 +12,8 @@ public class TiendaResponse {
     private String descripcion;
     private String direccion;
     private String horarios;
-    private Long planId;
+    private Integer planId;
     private String planNombre;
-    private Long userId;
+    private Integer userId;
     private String userEmail;
 }
